53 Max Subarray
How: Initialize a current sum and a global sum and set them both to be
equal to the initial array value. For each iteration, update the current
sum to the current number if the current number is larger than the new
sum would be. If the current sum is ever larger than the global sum, set
the global sum to be equal to the current sum. At the end, return the
global max.